requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
asgiref==3.5.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
cloud-init==24.4.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
djangorestframework==3.15.1
exceptiongroup==1.3.0
feedparser==6.0.11
Flask==3.1.1
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
openai==1.84.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.11.5
pydantic_core==2.33.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMySQL==1.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dotenv==1.1.0
pytz==2022.1
PyYAML==5.4.1
requests==2.32.4
SecretStorage==3.3.1
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.2
systemd-python==234
tqdm==4.67.1
Twisted==22.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.1.3
zipp==1.0.0
zope.interface==5.4.0
pytest
